Features/mehnaz/contact (PR #6 changes) : create ContactForm component and also remove @mui/icons-material package
