+++ package.json
@@ -13,7 +13,6 @@
     "@emotion/react": "^11.13.3",
     "@emotion/styled": "^11.13.0",
     "@hookform/resolvers": "^3.9.0",
-    "@mui/icons-material": "^6.1.3",
     "@mui/material": "^6.1.2",
     "@next/font": "^14.2.15",
     "next": "14.2.14",
